Add support to query for fetching a single record
Including nested nested has many relationship for a single record can
still produce N+1 queries so show actions which support that need to use
the query method as well. Added find, and find_by, which delegate to the
active record's find_by method, and added a record method, which is
intended to be used without parameters.